Chatbots/AzureFoundry/requirements.txt
Flask==3.1.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.64.0
httpx==0.27.0
httpx-sse==0.4.0
langchain-openai==0.3.7